MCQGenerator/requirements.txt
openai
langchain
langchain-community
huggingface_hub
langchain_huggingface
streamlit
python-dotenv
PyPDF2
transformers

-e .